Move some feature specific files into feature folder and rename
